Do not implicitly convert signal value to bool
On upcoming cocotb 2.0 such conversion is deprecated and will eventually
always return True. The recommended replacement is to convert to string
and then compare.

On cocotb 1.9 explicit conversion to string is no longer necessary. Once
this is minimum version of cocotb supported by cocotbext-axi, the
explicit conversions can be dropped.