Re-add support for event data field that has been removed in cocotb 2.0
The .data field of cocotb.triggers.Event has been removed in cocotb 2.0.
This is a part of the API exposed by cocotbext-axi. It is probably best
to keep the current API until support for cocotb 1.x has been removed
from cocotbext-axi.

Accordingly, the simpliest solution for this problem is to detect cocotb
2.0 version and re-add the .data field when event.set is being called on
cocotb 2.0.